+++ test/asphalt/core_test.clj
@@ -51,6 +51,8 @@ VALUES (^string $name, ^int $salary, ^string $dept, ^date $joined)")
 -- ^boolean gender,
 ^int salary, ^string dept, ^date j_date FROM emp")
 
+(a/defsql t-crosstype "SELECT ^string j_date FROM emp")
+
 (a/defsql t-selfew "SELECT ^string name, ^int salary, ^string dept, ^date j_date FROM emp WHERE name = ?")
 
 (a/defsql t-qfetch "SELECT ^string name, ^int salary, ^string dept, ^date j_date FROM emp WHERE name = ?"
@@ -64,6 +66,19 @@ VALUES (^string $name, ^int $salary, ^string $dept, ^date $joined)")
 
 
 ;; ----- tests -----
+
+
+(deftest test-crosstype
+  (let [jd1 (u/make-date)
+        vs1 ["Joe Coder" 100000 "Accounts" jd1]
+        row (zipmap [:name :salary :dept :joined] vs1)]
+    (let [generated-key (a/genkey u/ds t-insert row)]
+      (is (= 1 generated-key) "Verify that insertion generated a key"))
+    (is (= 1 (a/query a/fetch-single-value
+               u/ds t-count [])) "Verify that row was inserted")
+    (is (= (str jd1) (a/query a/fetch-single-value u/ds t-crosstype [])))
+    (is (= [(str jd1)] (a/query a/fetch-single-row u/ds t-crosstype [])))
+    (is (= [[(str jd1)]] (a/query a/fetch-rows u/ds t-crosstype [])))))
 
 
 (deftest test-params